third_party/blink/web_tests/platform/generic/fullscreen/full-screen-not-enabled-when-unsupported-expected.txt

This tests that the fullscreenEnabled property is false when fullscreen is not supported as laid out in section 4 of the Fullscreen API living standard
EXPECTED (document.webkitFullscreenEnabled == 'false') OK
EXPECTED (iframe.contentDocument.webkitFullscreenEnabled == 'false') OK
END OF TEST
